+++ CHANGES.md
@@ -6,3 +6,5 @@
 - Removed LT and LV modules in order to focus on Estonian (and generic) banklink specs. 
 
 - Removed Eclipse IDE files from the project.
+
+- Removed `Version` holder class.